Move communication zone to own component
