+++ src/app/components/communication-zone/communication-zone.component.sass
@@ -1,0 +1,11 @@
+@import ../../color
+
+.communication-container
+  display: flex
+  flex-direction: column
+  align-items: center
+
+.communication-area
+  height: 135px
+  width: 86px
+  background-color: $purple
