+++ src/app/components/player-summary/player-summary.component.sass
@@ -1,5 +1,3 @@
-@import ../../color
-
 .player-summary
   min-width: 65px
 
@@ -17,21 +15,5 @@
   max-width: 350px
   width: 35vw
 
-.interaction-container
-  display: flex
-  flex-direction: column
-  justify-content: space-around
-  max-height: 20em
-
 .card-summary-container
   display: flex
-
-.communication-container
-  display: flex
-  flex-direction: column
-  align-items: center
-
-.communication-area
-  height: 135px
-  width: 86px
-  background-color: $purple
