+++ src/app/components/player-summary/player-summary.component.spec.ts
@@ -1,11 +1,7 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
-import { CommunicationPositionPipe } from 'src/app/pipes/communication-position/communication-position.pipe';
 import { PlayerDisplayNamePipe } from '../../pipes/player-display-name/player-display-name.pipe';
 import { PlayerTaskListPipe } from '../../pipes/player-task-list/player-task-list.pipe';
 import { PlayerSummaryComponent } from './player-summary.component';
-
-const USERNAME_ONE = 'Custom name 1';
-const USERNAME_TWO = 'Custom name 2';
 
 describe('PlayerSummaryComponent', () => {
   let component: PlayerSummaryComponent;
@@ -13,11 +9,7 @@ describe('PlayerSummaryComponent', () => {
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      declarations: [
-        PlayerSummaryComponent,
-        PlayerTaskListPipe,
-        CommunicationPositionPipe,
-      ],
+      declarations: [PlayerSummaryComponent, PlayerTaskListPipe],
       providers: [PlayerDisplayNamePipe, PlayerTaskListPipe],
     }).compileComponents();
   });
@@ -30,33 +22,5 @@ describe('PlayerSummaryComponent', () => {
 
   it('should create', () => {
     expect(component).toBeTruthy();
-  });
-  it('should prevent dragging to players hand if they have communicated', () => {
-    const communicationCard = {
-      suit: 'pink',
-      value: 2,
-      playerPosition: 1,
-      username: USERNAME_TWO,
-      type: 'lowest',
-    };
-    component.communicationCard = [communicationCard];
-    component.cardCommunicated = true;
-    const expected = 'playing-mat';
-    const actual = component.communicationDragTo();
-    expect(actual).toBe(expected);
-  });
-
-  it('should allow dragging to players hand if they have not communicated', () => {
-    const communicationCard = {
-      suit: 'pink',
-      value: 2,
-      playerPosition: 1,
-      username: USERNAME_ONE,
-      type: 'lowest',
-    };
-    component.communicationCard = [communicationCard];
-    const expected = ['playing-mat', 'players-hand'];
-    const actual = component.communicationDragTo();
-    expect(actual).toEqual(expected);
   });
 });
